refactored strategy, instruments
