--- comet_pqc/driver/__init__.py
@@ -1,1 +1,0 @@
-from .e4980a import *
